refactor(sidebar): extract useSidebarTags hook
Move tag aggregation (readNote-with-cache over every unlocked note),
the stale-filter-clear effect, the per-note tag lookup, and the
selected-tags AND filter into a dedicated hook. Sidebar now just
consumes the two helpers it renders with.

Sidebar.tsx: 796 -> 740 LoC.
